+++ crates/xray-db/src/spawn_file/chunks/spawn_alife_spawns_chunk.rs
@@ -79,7 +79,7 @@ impl SpawnALifeSpawnsChunk {
 
       objects_writer.write_all(
         object_writer
-          .flush_chunk_into_buffer::<T>(index)?
+          .flush_chunk_into_buffer::<T>(index as u32)?
           .as_slice(),
       )?;
     }
